// UI/src/app/model/inspire-team.ts
export class InspireTeam
{
    InspireTeamId: number;
    teamName: string;
    Impact: string;
    BeASpark: string;
    VoterPayrollId: number;
    Period: string;
    Winner?: boolean;
    WinnerRationale?: string;
    DateCreated: Date;
    DateLastModified: Date;
    ModifiedBy?: string;
    EventId: number;
}